Implement Authorization using policies

+++ app/Http/Controllers/QuestionController.php
@@ -8,6 +8,10 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -15,11 +19,9 @@ class QuestionController extends Controller
      */
     public function index()
     {
-
-        $questions = Question::with('user')->latest()->paginate(5);
+        $questions = Question::with('user')->latest()->paginate(10);
         return view('questions.index', compact('questions'));
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -27,10 +29,9 @@ class QuestionController extends Controller
      */
     public function create()
     {
-        $questions = new Question();
-        return view('questions.create', compact($questions));
+        $question = new Question();
+        return view('questions.create', compact('question'));
     }
-
     /**
      * Store a newly created resource in storage.
      *
@@ -40,9 +41,8 @@ class QuestionController extends Controller
     public function store(AskQuestionRequest $request)
     {
         $request->user()->questions()->create($request->only('title', 'body'));
-        return redirect()->route('questions.index')->with('success', 'Your Question has been Posted Successfully');
+        return redirect()->route('questions.index')->with('success', "Your question has been submitted");
     }
-
     /**
      * Display the specified resource.
      *
@@ -54,7 +54,6 @@ class QuestionController extends Controller
         $question->increment('views');
         return view('questions.show', compact('question'));
     }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -63,9 +62,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        $this->authorize("update", $question);
+        return view("questions.edit", compact('question'));
     }
-
     /**
      * Update the specified resource in storage.
      *
@@ -75,10 +74,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question);
         $question->update($request->only('title', 'body'));
-        return redirect('/questions')->with('success', 'Your Question has been updated');
+        return redirect('/questions')->with('success', "Your question has been updated.");
     }
-
     /**
      * Remove the specified resource from storage.
      *
@@ -87,7 +86,8 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("delete", $question);
         $question->delete();
-        return redirect('/questions')->with('success', 'Question has been deleted');
+        return redirect('/questions')->with('success', "Your question has been deleted.");
     }
 }
